refactor: redirect provider imports to services, clean up shims
