+++ src/quodeq/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version as _pkg_version, PackageNotFoundError
 
-from quodeq.provider.base import ActionProvider
+from quodeq.services.base import ActionProvider
 
 try:
     __version__: str | None = _pkg_version("quodeq")
