+++ src/quodeq/provider/__init__.py
@@ -1,7 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services."""
-from quodeq.services import (  # noqa: F401
-    ActionProvider,
-    EvaluationOptions,
-    FilesystemActionProvider,
-    JobManager,
-)
+"""Legacy provider package — all implementations live in quodeq.services."""
