+++ src/quodeq/provider/accumulated.py
@@ -1,6 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.accumulated."""
-from quodeq.services.accumulated import (  # noqa: F401
-    AccumulatedCacheConfig,
-    compute_accumulated,
-    numeric_average,
-)
+"""Legacy shim — use quodeq.services.accumulated instead."""
