+++ src/quodeq/provider/base.py
@@ -1,9 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.base."""
-from quodeq.services.base import (  # noqa: F401
-    EvaluationOptions,
-    ProjectActions,
-    ReportActions,
-    EvaluationActions,
-    ToolingActions,
-    ActionProvider,
-)
+"""Legacy shim — use quodeq.services.base instead."""
