+++ src/quodeq/provider/dashboard.py
@@ -1,5 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.dashboard."""
-from quodeq.services.dashboard import (  # noqa: F401
-    DashboardCacheConfig,
-    build_dashboard,
-)
+"""Legacy shim — use quodeq.services.dashboard instead."""
