+++ src/quodeq/provider/evaluation_mixin.py
@@ -1,6 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.evaluation_mixin."""
-from quodeq.services.evaluation_mixin import (  # noqa: F401
-    EvaluationDispatcher,
-    SubprocessDispatcher,
-    FsEvaluationMixin,
-)
+"""Legacy shim — use quodeq.services.evaluation_mixin instead."""
