+++ src/quodeq/provider/filesystem.py
@@ -1,2 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.filesystem."""
-from quodeq.services.filesystem import FilesystemActionProvider  # noqa: F401
+"""Legacy shim — use quodeq.services.filesystem instead."""
