+++ src/quodeq/provider/tooling_mixin.py
@@ -1,5 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.tooling_mixin."""
-from quodeq.services.tooling_mixin import (  # noqa: F401
-    get_allowed_client_ids,
-    FsToolingMixin,
-)
+"""Legacy shim — use quodeq.services.tooling_mixin instead."""
