+++ src/quodeq/provider/violations.py
@@ -1,7 +1,1 @@
-"""Re-export for backward compatibility — moved to quodeq.services.violations."""
-from quodeq.services.violations import (  # noqa: F401
-    _ResolveOptions,
-    resolve_dimension_eval,
-    aggregate_violations,
-    ViolationContext,
-)
+"""Legacy shim — use quodeq.services.violations instead."""
